✨ (App.tsx): add support for auto login feature to improve user experience and reduce login friction 📝 (authContext.tsx): remove unused autoLogin function and refactor getUser function to improve code readability and maintainability 📝 (authContext.tsx): add getUser function to retrieve user data and improve user authentication flow 📝 (auth.ts): add getUser function to AuthContextType to improve type safety and provide a clear API for accessing user data
