+++ src/frontend/src/App.tsx
@@ -1,9 +1,8 @@
 import { useContext, useEffect, useState } from "react";
-import "reactflow/dist/style.css";
-import "./App.css";
-
 import { ErrorBoundary } from "react-error-boundary";
 import { useNavigate } from "react-router-dom";
+import "reactflow/dist/style.css";
+import "./App.css";
 import ErrorAlert from "./alerts/error";
 import NoticeAlert from "./alerts/notice";
 import SuccessAlert from "./alerts/success";
@@ -15,7 +14,7 @@ import {
   FETCH_ERROR_MESSAGE,
 } from "./constants/constants";
 import { AuthContext } from "./contexts/authContext";
-import { getGlobalVariables, getHealth } from "./controllers/API";
+import { autoLogin, getGlobalVariables, getHealth } from "./controllers/API";
 import Router from "./routes";
 import useAlertStore from "./stores/alertStore";
 import { useDarkStore } from "./stores/darkStore";
@@ -38,8 +37,10 @@ export default function App() {
     removeFromTempNotificationList(id);
   };
 
-  const { isAuthenticated } = useContext(AuthContext);
+  const { isAuthenticated, login, setUserData, setAutoLogin, getUser } =
+    useContext(AuthContext);
   const refreshFlows = useFlowsManagerStore((state) => state.refreshFlows);
+  const setLoading = useAlertStore((state) => state.setLoading);
   const fetchApiData = useStoreStore((state) => state.fetchApiData);
   const getTypes = useTypesStore((state) => state.getTypes);
   const refreshVersion = useDarkStore((state) => state.refreshVersion);
@@ -53,25 +54,45 @@ export default function App() {
   const [isLoadingHealth, setIsLoadingHealth] = useState(false);
 
   useEffect(() => {
-    refreshStars();
-    refreshVersion();
+    const isLoginPage = location.pathname.includes("login");
 
-    const fetchData = async () => {
-      if (isAuthenticated) {
-        try {
-          await getTypes();
-          refreshFlows();
-          const res = await getGlobalVariables();
-          setGlobalVariables(res);
-          checkHasStore();
-          fetchApiData();
-        } catch (error) {
-          console.error("Failed to fetch data:", error);
+    autoLogin()
+      .then(async (user) => {
+        if (user && user["access_token"]) {
+          user["refresh_token"] = "auto";
+          login(user["access_token"]);
+          setUserData(user);
+          setAutoLogin(true);
+          setLoading(false);
+          await Promise.all([refreshStars(), refreshVersion(), fetchData()]);
         }
-      }
-    };
-    fetchData();
+      })
+      .catch(async () => {
+        setAutoLogin(false);
+        if (isAuthenticated && !isLoginPage) {
+          getUser();
+          await Promise.all([refreshStars(), refreshVersion(), fetchData()]);
+        } else {
+          setLoading(false);
+          useFlowsManagerStore.setState({ isLoading: false });
+        }
+      });
   }, [isAuthenticated]);
+
+  const fetchData = async () => {
+    if (isAuthenticated) {
+      try {
+        await getTypes();
+        refreshFlows();
+        const res = await getGlobalVariables();
+        setGlobalVariables(res);
+        checkHasStore();
+        fetchApiData();
+      } catch (error) {
+        console.error("Failed to fetch data:", error);
+      }
+    }
+  };
 
   useEffect(() => {
     checkApplicationHealth();
