+++ src/frontend/src/types/contexts/auth.ts
@@ -15,4 +15,5 @@ export type AuthContextType = {
   apiKey: string | null;
   setApiKey: (apiKey: string | null) => void;
   storeApiKey: (apiKey: string) => void;
+  getUser: () => void;
 };
